requirements.txt
absl-py==0.13.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
black==21.4b2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.24
# Editable install with no version control (detectron2==0.5+cu102)
-e /home/PeterYuan/anaconda3/envs/detectron/lib/python3.7/site-packages
future==0.18.2
fvcore==0.1.5.post20210812
google-auth==1.35.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
hydra-core==1.1.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.6.4
importlib-resources==5.2.2
iopath==0.1.8
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.3
mypy-extensions==0.4.3
networkx==2.6.2
numpy==1.21.2
oauthlib==3.1.1
omegaconf==2.1.0
opencv-python==4.5.3.56
pathspec==0.9.0
Pillow==8.3.1
portalocker==2.3.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycococreatortools @ git+git://github.com/waspinator/pycococreator.git@fba8f4098f3c7aaa05fe119dc93bbe4063afdab8
pycocotools==2.0.2
pydot==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==5.1
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scipy==1.7.1
six==1.16.0
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
tifffile==2021.8.8
toml==0.10.2
torch==1.9.0+cu102
torchvision==0.10.0+cu102
tqdm==4.62.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
yacs==0.1.8
zipp==3.5.0
